pyproject.toml
[project]
name = "exercise-3"
version = "0.1.0"
description = ""
authors = [
    {name = "Moueed Ali",email = "moueedAli@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "numpy (>=2.4.0,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "fastapi[standard] (>=0.127.0,<0.128.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pymongo[srv] (>=4.15.5,<5.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "python-weather (>=2.1.1,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
